==> 1_preprocessing.ipynb <==
# Chuẩn bị dữ liệu & xử lý đặc trưng cho Titanic


## Bước 1: Load dữ liệu

import pandas as pd

train = pd.read_csv("train.csv")
test = pd.read_csv("test.csv")

print(train.shape)
train.head()


## Bước 2: Tạo đặc trưng mới


train['FamilySize'] = train['SibSp'] + train['Parch'] + 1
train['IsAlone'] = (train['FamilySize'] == 1).astype(int)

# Trích xuất danh xưng (Title) từ tên
train['Title'] = train['Name'].str.extract(' ([A-Za-z]+)\.', expand=False)
train['Title'] = train['Title'].replace(['Mlle','Ms'],'Miss').replace('Mme','Mrs')


## Bước 3: Chọn các cột cần thiết

features = ['Pclass','Sex','Age','Fare','Embarked','FamilySize','IsAlone','Title']
X = train[features]
y = train['Survived']


## Bước 4: Tiền xử lý dữ liệu

from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.impute import SimpleImputer

num_cols = ['Age', 'Fare', 'FamilySize']
cat_cols = ['Sex', 'Embarked', 'Pclass', 'Title']

num_pipe = Pipeline([
    ('imputer', SimpleImputer(strategy='median')),
    ('scaler', StandardScaler())
])

cat_pipe = Pipeline([
    ('imputer', SimpleImputer(strategy='most_frequent')),
    ('encoder', OneHotEncoder(handle_unknown='ignore'))
])

preprocessor = ColumnTransformer([
    ('num', num_pipe, num_cols),
    ('cat', cat_pipe, cat_cols)
])


## Bước 5: Fit và lưu pipeline

import joblib

preprocessor.fit(X)
joblib.dump(preprocessor, 'preprocessor.pkl')
print("Saved preprocessor.pkl")


## Bước 6: Áp dụng xử lý


X_ready = preprocessor.transform(X)
