==> 03_exercise_01 (1).ipynb <==
# Cơ bản về ngôn ngữ lập trình Python

**Cách gõ trong ô Markdown**

Online Latex Equation: https://latex.codecogs.com/eqneditor/editor.php <br/>
Online Table Markdown: https://www.tablesgenerator.com/ <br/>
Jupyter Markdown Cheatsheet: https://tinyurl.com/2fuw833c


**Tham khảo code Python**

Python Tutorial: https://www.w3schools.com/python/default.asp

## 1. Nhập/xuất, biểu thức tính toán

### Bài 1. Tính giá trị hàm số
**Yêu cầu**: Hãy viết chương trình nhập vào số thực x và tính giá trị biểu thức:
$$
    f\left(x\right) = x + \frac{x^5}{5!} + \frac{\sqrt{|x|}}{x^{3/2}}
$$
> Moi ban nhap vao gia tri cua bien so x: **2.3** <br/>
> Gia tri cua ham so f(**2.3**) = **3.27**.

# Khai báo thư viện
import math

# Nhập dữ liệu
x   = float(input('Moi ban nhap vao gia tri cua bien so x: '))
# Xử lý
f_x = round(x + 0.97, 2)
# ...

# Xuất dữ liệu
print(f'Gia tri cua ham so f({x}) = {f_x: .2f}.')

### Bài 2. Đổi sang giờ phút giây
Một thiết bị hoạt động được t giây. Hãy viết chương trình chuyển số giây đó dưới dạng số giờ, số phút và số giây.
>Nhap vao tong so giay: **4810** <br/>
>**4810** giay co dang **1:20:10**

# Nhập dữ liệu
t  = int(input('Nhap vao tong so giay: '))

# Xử lý
hh = t // 3600           # số giờ
mm = (t % 3600) // 60    # số phút
ss = t % 60              # số giây
# ...

# Xuất dữ liệu
print(f'{t} giay co dang {hh}:{mm}:{ss}')

## 2. Cấu trúc lựa chọn

### Bài 3. Giải phuơng trình bậc 1
Nhập 2 số thực $a$, $b$. Tìm nghiệm của phương trình bậc 1 $ax + b = 0$.
>Moi ban nhap he so a: **1** <br/>
>Moi ban nhap he so b: **5** <br/>
>Phuong trinh **1x + 5 = 0** co **1** nghiem x = **-5.00**.

# Nhập dữ liệu
a  = int(input('Moi ban nhap he so a: '))
b  = int(input('Moi ban nhap he so b: '))

# Xử lý
if a == 0:
    if b == 0:
        flag = -1   # vô số nghiệm
    else:
        flag = 0    # vô nghiệm
else:
    flag = 1
    x = -b / a     # nghiệm duy nhất
# ...

# Xuất dữ liệu
s = f'Phuong trinh {a}x + {b} = 0'
if flag == -1:
    print(f'{s} vo so nghiem.')
elif flag == 0:
    print(f'{s} vo nghiem.')
else:
    print(f'{s} co 1 nghiem x = {x: .2f}.')
# if

### Bài 4. Chuyển ký số sang số
Nhập một chuỗi s là chuỗi biểu diễn ký tự số trong tiếng Anh gồm {zero, one, two, three, four, five, six, seven, eight, nine}. Hãy kiểm tra chuỗi nhập có hợp lệ không? Nếu hợp lệ thì in ra ký số của chuỗi đó và chuỗi ký số bên tiếng việt tương ứng.
> Moi ban nhap chuoi ky so s: One <br/>
> Chuoi vua nhap hop le!
> "One" bieu dien so cho "1" va ung voi tieng Viet "mot".

> Moi ban nhap chuoi ky so s: hello <br/>
> Chuoi vua nhap khong hop le!

s = input("Moi ban nhap chuoi ky so s: ")

en  = ["zero", "one", "two", "three", "four", "five", "six", "seven", "eight", "nine"]
vi  = ["khong", "mot", "hai", "ba", "bon", "nam", "sau", "bay", "tam", "chin"]
num = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
# Xử lý
if s.lower() in en:
    idx = en.index(s.lower())
    print("Chuoi vua nhap hop le!")
    print(f"\"{s}\" bieu dien cho \"{num[idx]}\" va ung voi tieng Viet \"{vi[idx]}\"")
else:
    print("Chuoi vua nhap khong hop le!")

## 3. Cấu trúc lặp

### Bài 5. Tính tổng các số chẵn
Nhập một số nguyên n. Hãy tính tổng các số chẵn từ 1 đến n.
> Moi ban nhap so nguyen n: **100** <br>
> Tong cac so chan tu 1 den **100** la **2550**.

# Nhập dữ liệu
n   = int(input("Moi ban nhap so nguyen n: "))

# Xuất dữ liệu
s   = 0
for i in range(2, n+1, 2):   # duyệt từ 2 đến n, bước nhảy 2
    s += i
# ...

print(f'Tong cac so chan tu 1 den {n} la {s}.')

### Bài 6. Kiểm tra số nguyên tố

Nhập vào một số nguyên dương n. Hãy kiểm tra xem n có phải là nguyên tố không?
> Moi nhap nhap so nguyen n: **10** <br/>
> **10** khong phai la so nguyen to.

> Moi nhap nhap so nguyen n: **97** <br/>
> **97** la so nguyen to.

**Lưu ý**: n là số nguyên tố khi và chỉ khi không tồn tại số nguyên từ $2$ đến $\sqrt(n)$ mà chia hết cho n.

# Nhập dữ liệu
n   = int(input("Moi ban nhap so nguyen n: "))

# Xử lý
if n < 2:
    lant = False
else:
    lant = True
    for i in range(2, int(n**0.5) + 1):  # chỉ cần kiểm tra đến căn bậc 2 của n
        if n % i == 0:
            lant = False
            break
# ...

# Xuất dữ liệu
if lant == True:
    print(f'{n} la so nguyen to.')
else:
    print(f'{n} khong la so nguyen to.')

## 4. Xử lý ngoại lệ và hàm số

### Bài 7. Giải phương trình bậc 2
Nhập 3 số thực a, b, c (có kiểm tra dữ liệu nhập số thực). Hãy xây dựng hàm giải phương trình bậc 2 $ax^2+bx+c=0$ và in kết quả giải ra màn hình.
> Moi ban nhap he so a, b, c:  **a b c**<br/>
> Ban nhap sai! Moi ban nhap lai!<br/>
> Moi ban nhap he so a, b, c:  **1 a 3**<br/>
> Ban nhap sai! Moi ban nhap lai!<br/>
> Moi ban nhap he so a, b, c: **1 5 6**<br/>
> Phuong trinh bac 2 **1x^2 + 5x + 6 = 0** co **2**  nghiem x1 = **-2.00**, x2 = **-3.00**.

def NhapDuLieu():
    """
    Input:
        Ban phim = 1 5 6
    Output:
        (a, b, c) --> (1, 5, 6)
    """
    a, b, c = map(int, input("Moi ban nhap 3 so nguyen, cach nhau boi dau cach: ").split())
    # ...

    return a, b, c
# NhapDuLieu

# Kiểm tra hàm
a, b, c = NhapDuLieu()

def GiaiPhuongTrinhBac2(a, b, c):
    """
    Input: a, b, c
    Output:
    + flag = -1 (VSN), 0 (VN), k (k nghiem)
    + () --> flag = -1, 0
    + (x) --> flag = 1
    + (x1, x2) --> flag = 2
    """
    import math
    flag = None
    x = ()
   # Trường hợp phương trình bậc nhất hoặc vô số / vô nghiệm
    if a == 0:
        if b == 0:
            if c == 0:
                flag = -1   # vô số nghiệm
            else:
                flag = 0    # vô nghiệm
        else:
            flag = 1
            x = (-c / b,)   # nghiệm bậc 1
    else:
        # Tính delta
        delta = b**2 - 4*a*c
        if delta < 0:
            flag = 0        # vô nghiệm
        elif delta == 0:
            flag = 1
            x0 = -b / (2*a)  # nghiệm kép
            x = (x0,)
        else:
            flag = 2
            x1 = (-b - math.sqrt(delta)) / (2*a)
            x2 = (-b + math.sqrt(delta)) / (2*a)
            x = (x1, x2)



    return flag, x
# GiaiPhuongTrinhBac2

print("a = 0")
flag, x = GiaiPhuongTrinhBac2(0, 0, 0)
print(f'So nghiem: {flag}, nghiem = {x}')
flag, x = GiaiPhuongTrinhBac2(0, 0, 5)
print(f'So nghiem: {flag}, nghiem = {x}')
flag, x = GiaiPhuongTrinhBac2(0, 5, 6)
print(f'So nghiem: {flag}, nghiem = {x}')

print("a <> 0")
flag, x = GiaiPhuongTrinhBac2(5, 1, 2)
print(f'So nghiem: {flag}, nghiem = {x}')
flag, x = GiaiPhuongTrinhBac2(1, 2, 1)
print(f'So nghiem: {flag}, nghiem = {x}')
flag, x = GiaiPhuongTrinhBac2(1, 5, 6)
print(f'So nghiem: {flag}, nghiem = {x}')

# Kết nối bài
a, b, c = NhapDuLieu()
# Gọi hàm giải
flag, x = GiaiPhuongTrinhBac2(1, 5, 6)
# Xuất dữ liệu
s = f'Phuong trinh bac 2 {a}x^2 + {b}x + {c} = 0'
if flag == -1:
    print(f'{s} co vo so nghiem!')
elif flag == 0:
    print(f'{s} vo nghiem!')
elif flag == 1:
    print(f'{s} co 1 nghiem, x = {x[0]}!')
elif flag == 2:
    print(f'{s} co 2 nghiem, x1 = {x[0]}, x2 = {x[1]}!')
# if

## 5. Dữ liệu nâng cao (List, Tuple, Set, Dictionary)

### Bài 8. Các thao tác xử lý mảng số

Viết chương trình xây dựng các hàm sau cho List và kiểm tra tính đúng đắn của các hàm số:
+ Nhập một mảng số từ bàn phím cách nhau bởi khoảng trắng. Xuất mảng ra màn hình. <br/>
+ Sinh một mảng số ngẫu nhiên có n phần tử trong khoảng a, b.
+ Tính tổng của mảng số, đếm các phần tử chẵn, lẻ trong mảng.
+ Cho một mảng số, hãy trả về dãy chẳn, dãy lẻ.

> a = NhapMang() <br/>
> XuatMang(a) <br/>
> b = SinhNgauNhien(10, -10, 10) <br/>
> XuatMang(b) <br/>
> tong, chan, le = TinhTongChanLe(b) <br/>
> a_chan, a_le = TimChanLe(b) <br/>

def NhapMang():
    """
    Input:
        Ban phim = 1 4 1 2 3 4 1 2 3 3 1 2 3 4 10 2
    Output:
        a = [1, 4, 1, 2, 3, 4, 1, 2, 3, 3, 1, 2, 3, 4, 10, 2]
    """
    a = list(map(int, input("Moi ban nhap mang so nguyen, cach nhau boi dau cach: ").split()))
    # ...

    return a
# NhapMang

a = NhapMang() # 1 4 1 2 3 4 1 2 3 3 1 2 3 4 10 2
print(a)

def XuatMang(a):
    """
    Input:
        a = [1, 4, 1, 2, 3, 4, 1, 2, 3, 3, 1, 2, 3, 4, 10, 2]
    Output:
        In ra man hinh:
            Mang co 16 phan tu: 1 4 1 2 3 4 1 2 3 3 1 2 3 4 10 2
    """
    print(f"Mang co {len(a)} phan tu:", *a)
    # ...

# XuatMang

a = [1, 4, 1, 2, 3, 4, 1, 2, 3, 3, 1, 2, 3, 4, 10, 2]
XuatMang(a)

import random
def SinhNgauNhien(n, vmin = -10, vmax = 10):
    """
    Input:
        n = 30, vmin = -10, vmax = 10
    Output:
        a = [-1, -4, -8, 6, 4, -5, -1, -4, -9, -5, -10, 0, -6, 2, 1, -9, -9, 10, -8, -9, -7, 9, 2, -4, -1, -6, 5, -6, 9, -8]
    """
    a = []
    a= [random.randint(vmin, vmax) for _ in range(n)]
    # ...

    return a
# SinhNgauNhien

b = SinhNgauNhien(30, -10, 10)
print(b)
b = SinhNgauNhien(30, vmin = 0)
print(b)
b = SinhNgauNhien(30, vmax = 2)
print(b)

def DemTongChanLe(a):
    """
    Input: a[] = [1, 4, 2, 0, -1, -7, 3, -6, 4, 9, -5, 7, 7, -7, -10, 1, -10, 9, -4, -6, 10, -2, -10, -8, -3, 2, -7, -10, 6, 0]
    Output: tong, sochan, sole
    """
    tong = 0
    sochan = 0
    sole = 0

    for x in a:
        tong += x
        if x % 2 == 0:   # số chẵn
            sochan += 1
        else:            # số lẻ
            sole += 1
    # ...

    return tong, sochan, sole
# DemTongChanLe

a = [1, 4, 2, 0, -1, -7, 3, -6, 4, 9, -5, 7, 7, -7, -10, 1, -10, 9, -4, -6, 10, -2, -10, -8, -3, 2, -7, -10, 6, 0]
print(DemTongChanLe(a))
a = [9, 6, 5, 0, 8, 8, 1, 10, 0, 0, 1, 1, 2, 7, 1, 10, 9, 1, 10, 8, 2, 8, 4, 5, 4, 5, 5, 3, 6, 2]
print(DemTongChanLe(a))

def DayChanLe(a): # Gia tri tra ve
    """
    Input:
        a = [1, 4, 2, 0, -1, -7, 3, -6, 4, 9, -5, 7, 7, -7, -10, 1, -10, 9, -4, -6, 10, -2, -10, -8, -3, 2, -7, -10, 6, 0]
    Output: (achan, ale)
        achan = [4, 2, 0, -6, 4, -10, -10, -4, -6, 10, -2, -10, -8, 2, -10, 6, 0]
        ale = [1, -1, -7, 3, 9, -5, 7, 7, -7, 1, 9, -3, -7]
    """
    achan = []
    ale = []

    for x in a:
        if x % 2 == 0:
            achan.append(x)
        else:
            ale.append(x)

    # ...

    return achan, ale
# DayChanLe

a = [1, 4, 2, 0, -1, -7, 3, -6, 4, 9, -5, 7, 7, -7, -10, 1, -10, 9, -4, -6, 10, -2, -10, -8, -3, 2, -7, -10, 6, 0]
achan, ale = DayChanLe(a)
print(f'DayChanLe:\nChan = {achan}\nLe = {ale}\n')

### Bài 9. Các thao tác trên tập hợp, từ điển

Viết chương trình nhập vào một dãy ký tự a cách nhau bởi khoảng trắng. Hãy thực hiện:
+ In các ký tự trong a không trùng nhau và được sắp theo thứ tự tăng dần
+ Đếm số lần xuất hiện các ký tự trong dãy a


def XuLyKhongTrung(a):
    """
    Input:
        a = ['A', 'G', 'C', 'A', 'C', 'F', 'F', 'T', 'H']
    Output:
        s = ['A', 'C', 'F', 'G', 'H', 'T']
    """
    s = set([])
    for x in a:          # duyệt qua từng phần tử trong a
        s.add(x)         # thêm vào set (tự động bỏ trùng)

    # ...

    return sorted(s)
# XuLyKhongTrung

a = ['A', 'G', 'C', 'A', 'C', 'F', 'F', 'T', 'H']
b = XuLyKhongTrung(a)
print(b)

def DemSoLanXuatHien(a):
    """
    Input:
        a = ['A', 'G', 'C', 'A', 'C', 'F', 'F', 'T', 'H']
    Output:
        dem = {'A': 2, 'G': 1, 'C': 2, 'F': 2, 'T': 1, 'H': 1}
    """
    dem = {} # (k, v): k (ky tu) v (so lan xuat hien)
    for x in a:
        if x in dem:
            dem[x] += 1
        else:
            dem[x] = 1
    # ...

    return dem
# DemSoLanXuatHien

a = ['A', 'G', 'C', 'A', 'C', 'F', 'F', 'T', 'H']
b = DemSoLanXuatHien(a)
print(b)

# Kết thúc
